% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histoires_datagochi_usa.R
\name{add_datagotchi_logo}
\alias{add_datagotchi_logo}
\title{Add Datagotchi Logo to a ggplot}
\usage{
add_datagotchi_logo(
  plot,
  logo_url = "https://raw.githubusercontent.com/clessn/img/refs/heads/main/Logo.PNG",
  logo_width = 0.1,
  x_pos = 1,
  y_pos = 0,
  hjust = 1,
  vjust = 0
)
}
\arguments{
\item{plot}{A `ggplot` object. The plot to which the logo will be added.}

\item{logo_url}{A character string specifying the URL of the logo image. 
Defaults to "https://raw.githubusercontent.com/clessn/img/refs/heads/main/Logo.PNG".}

\item{logo_width}{A numeric value specifying the width of the logo. Defaults to 0.1.}

\item{x_pos}{A numeric value for the x position of the logo. Defaults to 1 (right edge of the plot).}

\item{y_pos}{A numeric value for the y position of the logo. Defaults to 0 (bottom edge of the plot).}

\item{hjust}{Horizontal justification of the logo. Defaults to 1 (right-aligned).}

\item{vjust}{Vertical justification of the logo. Defaults to 0 (bottom-aligned).}
}
\value{
A `ggplot` object with the logo added.
}
\description{
This function overlays the Datagotchi logo onto a ggplot using the `cowplot` package.
The logo is placed as an image on top of the existing plot.
}
\examples{
library(ggplot2)
library(cowplot)
p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
add_datagotchi_logo(p)

# Adjust the logo position and size
add_datagotchi_logo(p, logo_width = 0.15, x_pos = 0.95, y_pos = 0.05)
}
