% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histoires_datagochi_usa.R
\name{check_fonts_access}
\alias{check_fonts_access}
\title{Check if font files are accessible}
\usage{
check_fonts_access()
}
\value{
A message indicating whether each font file is found or not.
}
\description{
This function checks if the specified font files are accessible via system.file().
}
