% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histoires_datagochi_usa.R
\docType{data}
\name{datagotchi_green_dark}
\alias{datagotchi_green_dark}
\title{Datagotchi Green Dark}
\format{
An object of class \code{character} of length 1.
}
\usage{
datagotchi_green_dark
}
\value{
A character string representing the color code "#26E92E".
}
\description{
A color code for the Datagotchi theme.
}
\keyword{datasets}
