% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histoires_datagochi_usa.R
\docType{data}
\name{datagotchi_green_light}
\alias{datagotchi_green_light}
\title{Datagotchi Green Light}
\format{
An object of class \code{character} of length 1.
}
\usage{
datagotchi_green_light
}
\value{
A character string representing the color code "#1BF640".
}
\description{
A color code for the Datagotchi theme.
}
\keyword{datasets}
