% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clessn_decision_tree.R
\name{draw_datagotchi_node}
\alias{draw_datagotchi_node}
\title{Draw Datagotchi Node}
\usage{
draw_datagotchi_node(
  pred_democrat,
  pred_independant,
  pred_republican,
  node_proportion,
  label,
  left_lab,
  right_lab
)
}
\arguments{
\item{pred_democrat}{Numeric. The proportion of predictions classified as Democrat in this node (between 0 and 1).}

\item{pred_independant}{Numeric. The proportion of predictions classified as Independent in this node (between 0 and 1).}

\item{pred_republican}{Numeric. The proportion of predictions classified as Republican in this node (between 0 and 1).}

\item{node_proportion}{Numeric. The proportion of the total sample represented by this node (between 0 and 1).}

\item{label}{Character. The label for the node, typically representing the decision criterion.}

\item{left_lab}{Character. The label to be displayed on the left side of the node (e.g., corresponding to the "Yes" branch).}

\item{right_lab}{Character. The label to be displayed on the right side of the node (e.g., corresponding to the "No" branch).}
}
\value{
A ggplot object representing the graphical node.
}
\description{
This function creates a graphical representation of a decision tree node, illustrating the proportion of predicted outcomes for Democrat, Independent, and Republican categories, along with the proportion of the total sample in that node.
}
\examples{
draw_datagotchi_node(
  pred_democrat = 0.5,
  pred_independant = 0.2,
  pred_republican = 0.3,
  node_proportion = 0.7,
  label = "Prius or Pickup?",
  left_lab = "Prius",
  right_lab = "Pickup"
)

}
