% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clessn_decision_tree.R
\name{draw_node}
\alias{draw_node}
\title{Draw a Node for Decision Tree Visualization}
\usage{
draw_node()
}
\value{
A plot representing a node in the decision tree.
}
\description{
This is a temporary function used to plot nodes in a larger decision tree visualization. 
The function currently plots a basic scatter plot as a placeholder.
}
