% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggboule.R
\name{jml}
\alias{jml}
\title{Print ASCII portrait of JML}
\usage{
jml()
}
\value{
No return value, called for side effect (printing ASCII art)
}
\description{
This function prints an ASCII art portrait of our boss that fits nicely in the RStudio console.
}
\examples{
jml()
}
