% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histoires_datagochi_usa.R
\docType{data}
\name{party_colors}
\alias{party_colors}
\title{Party Colors}
\format{
A named character vector with color codes.
}
\usage{
party_colors
}
\description{
A named vector containing the colors associated with the USA political parties (2024).
}
\keyword{datasets}
