% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartogram.R
\docType{data}
\name{state_coordinates}
\alias{state_coordinates}
\title{State Coordinates Cartogram}
\format{
A list of length 50 where each element is a named vector with two numeric values: `x` and `y`.
}
\usage{
state_coordinates
}
\description{
A list containing the x and y coordinates for each U.S. state to create a cartogram.
}
\keyword{datasets}
