% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggboule.R
\name{ggboule}
\alias{ggboule}
\title{Create a ggplot with image overlay on bars}
\usage{
ggboule(
  data,
  mapping = aes(),
  image_path = system.file("extdata/boule.png", package = "clessnize"),
  ...
)
}
\arguments{
\item{data}{The data frame to use for the plot}

\item{mapping}{Default list of aesthetic mappings to use for plot}

\item{image_path}{The path to the image to overlay on bars}

\item{...}{Other arguments passed to ggplot}
}
\value{
A ggplot object with the image overlay on bars
}
\description{
Create a ggplot with image overlay on bars
}
